+++ extensions/error.py
@@ -7,23 +7,23 @@ plugin = lightbulb.Plugin('error', 'error handlers for the commands.')
 @plugin.listener(lightbulb.CommandErrorEvent)
 async def on_error(event: lightbulb.CommandErrorEvent):
     if isinstance(event.exception, lightbulb.CommandInvocationError):
-        await event.context.respond(f"Something went wrong during invocation of command `{event.context.command.name}`.")
+        await event.context.respond(f"bruh a dum error occured during invocation of command `{event.context.command.name}`... check terminal!")
         raise event.exception
     exception = event.exception.__cause__ or event.exception
     if isinstance(exception, lightbulb.NotOwner):
-        await event.context.respond("You are not the owner of this bot.")
+        await event.context.respond("you aint my owner")
     elif isinstance(exception, lightbulb.CommandIsOnCooldown):
-        await event.context.respond(f"This command is on cooldown. Retry in `{exception.retry_after:.2f}` seconds.")
+        await event.context.respond(f"slowdown... retry in `{exception.retry_after:.2f}` seconds!")
     elif isinstance(exception, lightbulb.NotEnoughArguments):
-        await event.context.respond("Not enough arguments in user's input.")
+        await event.context.respond("be more specific")
     elif isinstance(exception, lightbulb.NSFWChannelOnly):
-        await event.context.respond("This command is restricted to only being used in NSFW channels.")
-    elif isinstance(exception, lightbulb.BotOnly):
-        await event.context.respond("This command is restricted to only being used by Bots.")
+        await event.context.respond("this command is too LEWD for this channel (;𖦹ㅁ𖦹)")
+    elif isinstance(exception, lightbulb.CommandNotFound):
+        await event.context.respond("i searched the whole universe and still couldn't find that command")
     elif isinstance(exception, lightbulb.BotMissingRequiredPermission):
-        await event.context.respond("Bot missing one or more permissions required to run this command.")
+        await event.context.respond("gimme perms first")
     elif isinstance(exception, lightbulb.MissingRequiredRole) or isinstance(exception, lightbulb.MissingRequiredPermission):
-        await event.context.respond("User missing role or permission required to run this command.")
+        await event.context.respond("u dont have the required perms or role")
     else:
         raise exception